// Practise/src/app/models/teacherReq.ts
export class TeacherReq
{

    name: string="";
    email: string="";
    phonenumber: string="";
    address: string="";
    dateOfBirth:string = "";
    genderId: number=0;
    countryId: number=0;
    stateId: number=0;
    cityId: number=0;
    departmentId: number=0;
    collegeId: number=0;
    profilepic:any;
}